migration , db seed

+++ database/migrations/2023_06_16_112632_create_products_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('name')->nullable(false); 
-            $table->bigInteger('price')->nullable(false); 
+            $table->float('price')->nullable(false); 
             $table->text('description')->nullable(true); 
             $table->enum('category', ['food' ,'toys','accessories','beds','grooming'])->nullable(false); 
             $table->text('image')->nullable(false);
